Former-commit-id: f932eee52a90a81a88b2ec584d45beb5a6251d06

+++ backend/src/main/java/com/ssafy/api/model/MyBadge.java
@@ -26,5 +26,5 @@ public class MyBadge {
 	private long userID;
 
 	@Column(name = "BADGE_ID")
-	private int badgeID;
+	private long badgeID;
 }
